Q26,27,28,29
Q26) count the total number of squares that can be visited by the Bishop in one move.
Q27) Find the position where the Ath item will be delivered if we start from a given position C in a circle of sizeB.
Q28) Given an even number ( greater than 2 ), return two prime numbers whose sum will be equal to given number.
Q29) find the sum of hamming distances of all pairs of integers in the array.