Q35&36&37
Q35) You are given an integer N and the task is to reverse the digits of the given integer. Return 0 if the result overflows and does not fit in a 32 bit signed integer.
Q36) Determine whether an integer is a palindrome(With String).
Q37) Determine whether an integer is a palindrome(Without String).