Q46 & 47
Q46) Given a sorted array of integers A(0 based index) of size N, find the starting and ending position of a given integer B in array A.
Q47) Given a sorted array A and a target value B, return the index if the target is found. If not, return the index where it would be if it were inserted in order.