Q80&81
Q80) Given an array and a value, remove all the instances of that value in the array.
Also return the number of elements left in the array after the operation.
Q81) Given a binary array A and a number B, we need to find length of the longest subsegment of ‘1’s possible by changing at most B ‘0’s.