Q91&92
Q91) The Linked List is binary i.e data values in the linked list nodes consist of only 0's and 1's.You need to sort the linked list and return the new linked list.
Q92) You need to find the value of the Bth node from the middle towards the beginning of the Linked List A.If no such element exists, then return -1.