Q104&105
Q104) You are given two linked lists representing two non-negative numbers. The digits are stored in reverse order and each of their nodes contain a single digit. Add the two numbers and return it as a linked list.
Q105) Detect And Remove Loop In A Linked List, Given a linked list, return the node where the cycle begins. If there is no cycle, return null.