Q111&112&113&114
Q111) Given a string A consisting only of '(' and ')'.You need to find whether parantheses in A is balanced or not
Q112)  Given a string A representing an absolute path for a file (Unix-style). Return the string A after simplifying the absolute path.
Q113) Given a string A denoting an expression. It contains the following operators '+', '-', '', '/'.Check whether A has redundant braces or not.
Q114) Design a stack that supports push, pop, top, and retrieving the minimum element in constant time.