Q117&118&119&120&121&122
Q117)find first non-repeating character each time a character is inserted to the stream and append it at the end to B. If no non-repeating character is found then append '#' at the end of B.
Q118)Find the area of the largest rectangle formed by the histogram.
Q119)find the maximum special product of any integer in the array.Product of LeftSpecialValue & RightSpecialValue.
Q120)Given an array, find the nearest smaller element G[i] for every element A[i] in the array such that the element has an index smaller than i.
Q121)Given an integer array A of non-negative integers representing an elevation map where the width of each bar is 1, compute how much water it is able to trap after raining.
Q122) There is a sliding window of size B which is moving from the very left of the array to the very right. You can only see the w numbers in the window. Each time the sliding window moves rightwards by one position. You have to find the maximum for each window.